% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl.R
\name{nhl_scores}
\alias{nhl_scores}
\title{Fetch NHL scores for today}
\usage{
nhl_scores(date = Sys.Date())
}
\arguments{
\item{date}{Date string "YYYY-MM-DD" (default: today)}
}
\value{
tibble: same as nhl_schedule
}
\description{
Fetch NHL scores for today
}
