% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl.R
\name{nhl_standings}
\alias{nhl_standings}
\title{Fetch current NHL standings}
\usage{
nhl_standings(date = Sys.Date())
}
\arguments{
\item{date}{Date string "YYYY-MM-DD" (default: today)}
}
\value{
tibble: team_abbrev, team_name, conference, division, games_played,
wins, losses, ot_losses, points, point_pctg, goal_for, goal_against,
goal_differential, streak_code, streak_count
}
\description{
Fetch current NHL standings
}
